Update Win_CyberCNS_Install.ps1
Moved tenant, portal and executablelocation to key store as they should never change client-to-client for 99% of users. Removed -Type entirely as a custom field and indicated it should be just passed as a variable. Didn't really understand why you'd want to specify this at the client level. Maybe agent level. But I suspect most people would pass this at the policy level (-Type Probe on a server and -Type LightWeight on a workstation, for example). Added a short explanation of the Uninstall variable.