Create linux_docker_run_commands.sh
This shell script is only for Linux machines running docker. It downloads an image and runs a small Docker container that pulses the run commands of every container in Docker. I find this useful for keeping track of the containers I am running in case I need to run them again.