Create linux_memory_check.sh
This script is an extension on the current Linux memory check script. I have modified it to include physical and swap. It also has 3 thresholds: OK, low and critically low.

It checks the free physical mem first. If the physical memory is less than min, the script will check swap.
If swap is more than ok amount, the script gives a warning.
If swap is less than ok amount, the script gives a warning.
If swap is also lower than min, the script gives an error.

The script allows the user to specify the threshold using ENV VARs, but if they don't, it uses defaults.